+++ vitest.config.ts
@@ -5,7 +5,10 @@ import viteConfig from './vite.config.ts'
 export default mergeConfig(viteConfig, defineConfig({
   test: {
     environment: 'jsdom',
-    setupFiles: ['./src/__tests__/setup.ts'],
+    setupFiles: [
+      './server/__tests__/setup.ts',
+      './src/__tests__/setup.ts',
+    ],
     globals: true,
     include: ['src/**/*.{test,spec}.{ts,tsx}', 'server/**/__tests__/*.{test,spec}.ts'],
     exclude: ['node_modules', 'dist', '.next'],